+++ src/common/js/global.js
@@ -2,12 +2,12 @@
  * pwix:options/src/common/js/global.js
  */
 
-import { Options } from '../classes/options.class.js';
+import { BaseOpt } from '../classes/base.class.js';
 
 pwixOptions = {
     // package configuration
     _conf: {},
 
     // the exported class
-    Options: Options
+    BaseOpt: BaseOpt
 };